pyproject.toml
[tool.poetry]
name = "djangocms-animate-on-scroll"
version = "1.0"
description = ""
authors = ["michalsnik, Michael Carder", "Fabien Michel <fabien.michel@hespul.org>"]

[tool.poetry.dependencies]
python = ">=3.8"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=2.20"
black = ">=22.12"
isort = ">=5.11"
pylint = ">=2.15"
pylint-django = ">=2.5"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
atomic = true
profile = "black"
line_length = 120
known_django = "django"
# known_thirdparty = "strawberry_django_plus, strawberry"
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

[tool.djlint]
profile = "django"
max_line_length = 120
indent = 2
blank_line_after_tag = "load,extends,include"

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "invalid-name",
    "arguments-differ",
    "logging-fstring-interpolation",
    "too-few-public-methods",
    "arguments-renamed",
    "wrong-import-order",
    "import-outside-toplevel",
    "no-member",
    "empty-docstring",
    "unused-argument",
    "abstract-method",
    "protected-access",
    "too-many-arguments",
    "broad-except",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"